use trait to decouple types
