+++ crates/cfxcore/pow/src/lib.rs
@@ -7,22 +7,23 @@ mod compute;
 mod keccak;
 mod seed_compute;
 mod shared;
+mod target_difficulty_manager;
+mod traits;
 
-pub use self::{cache::CacheBuilder, shared::POW_STAGE_LENGTH};
+pub use traits::ConsensusProvider;
+
+pub use self::{
+    cache::CacheBuilder, shared::POW_STAGE_LENGTH,
+    target_difficulty_manager::TargetDifficultyManager,
+};
 use keccak_hash::keccak as keccak_hash;
 
 use cfx_parameters::pow::*;
 use cfx_types::{BigEndianHash, H256, U256, U512};
-use malloc_size_of::{MallocSizeOf, MallocSizeOfOps};
+
 use malloc_size_of_derive::MallocSizeOf as DeriveMallocSizeOf;
-use parking_lot::RwLock;
-use primitives::BlockHeader;
 use static_assertions::_core::str::FromStr;
-use std::{
-    collections::{HashMap, VecDeque},
-    convert::TryFrom,
-    sync::Arc,
-};
+use std::{convert::TryFrom, sync::Arc};
 
 #[cfg(target_endian = "big")]
 compile_error!("The PoW implementation requires little-endian platform");
@@ -318,167 +319,6 @@ pub fn validate(
         &nonce,
         &problem.boundary,
     )
-}
-
-/// This function computes the target difficulty of the next period
-/// based on the current period. `cur_hash` should be the hash of
-/// the block at the current period upper boundary and it must have been
-/// inserted to BlockDataManager, otherwise the function will panic.
-/// `num_blocks_in_epoch` is a function that returns the epoch size
-/// under the epoch view of a given block.
-pub fn target_difficulty<F, M>(
-    block_header_by_hash: M,
-    target_difficulty_manager: &TargetDifficultyManager,
-    pow_config: &ProofOfWorkConfig, cur_hash: &H256, num_blocks_in_epoch: F,
-) -> U256
-where
-    F: Fn(&H256) -> usize,
-    M: Fn(&H256) -> Option<Arc<BlockHeader>>,
-{
-    if let Some(target_diff) = target_difficulty_manager.get(cur_hash) {
-        // The target difficulty of this period is already computed and cached.
-        return target_diff;
-    }
-
-    let mut cur_header = block_header_by_hash(cur_hash)
-        .expect("Must already in BlockDataManager block_header");
-    let epoch = cur_header.height();
-    assert_ne!(epoch, 0);
-    debug_assert!(
-        epoch
-            == (epoch / pow_config.difficulty_adjustment_epoch_period(epoch))
-                * pow_config.difficulty_adjustment_epoch_period(epoch)
-    );
-
-    let mut cur = cur_hash.clone();
-    let cur_difficulty = cur_header.difficulty().clone();
-    let mut block_count = 0 as u64;
-    let max_time = cur_header.timestamp();
-    let mut min_time = 0;
-
-    // Collect the total block count and the timespan in the current period
-    for _ in 0..pow_config.difficulty_adjustment_epoch_period(epoch) {
-        block_count += num_blocks_in_epoch(&cur) as u64;
-        cur = cur_header.parent_hash().clone();
-        cur_header = block_header_by_hash(&cur).unwrap();
-        if cur_header.timestamp() != 0 {
-            min_time = cur_header.timestamp();
-        }
-        assert!(max_time >= min_time);
-    }
-
-    let expected_diff = pow_config.target_difficulty(
-        block_count,
-        max_time - min_time,
-        &cur_difficulty,
-    );
-    // d_{t+1}=0.8*d_t+0.2*d'
-    // where d_t is the difficulty of the current period, and d' is the
-    // expected difficulty to reach the ideal block_generation_period.
-    let mut target_diff = if epoch < pow_config.cip86_height {
-        expected_diff
-    } else {
-        cur_difficulty / 5 * 4 + expected_diff / 5
-    };
-
-    let (lower, upper) = pow_config.get_adjustment_bound(cur_difficulty);
-    if target_diff > upper {
-        target_diff = upper;
-    }
-    if target_diff < lower {
-        target_diff = lower;
-    }
-
-    // Caching the computed target difficulty of this period.
-    target_difficulty_manager.set(*cur_hash, target_diff);
-
-    target_diff
-}
-
-//FIXME: make entries replaceable
-#[derive(DeriveMallocSizeOf)]
-struct TargetDifficultyCacheInner {
-    capacity: usize,
-    meta: VecDeque<H256>,
-    cache: HashMap<H256, U256>,
-}
-
-impl TargetDifficultyCacheInner {
-    pub fn new(capacity: usize) -> Self {
-        TargetDifficultyCacheInner {
-            capacity,
-            meta: Default::default(),
-            cache: Default::default(),
-        }
-    }
-
-    pub fn is_full(&self) -> bool { self.meta.len() >= self.capacity }
-
-    pub fn evict_one(&mut self) {
-        let hash = self.meta.pop_front();
-        if let Some(h) = hash {
-            self.cache.remove(&h);
-        }
-    }
-
-    pub fn insert(&mut self, hash: H256, difficulty: U256) {
-        self.meta.push_back(hash.clone());
-        self.cache.insert(hash, difficulty);
-    }
-}
-
-struct TargetDifficultyCache {
-    inner: RwLock<TargetDifficultyCacheInner>,
-}
-
-impl MallocSizeOf for TargetDifficultyCache {
-    fn size_of(&self, ops: &mut MallocSizeOfOps) -> usize {
-        self.inner.read().size_of(ops)
-    }
-}
-
-impl TargetDifficultyCache {
-    pub fn new(capacity: usize) -> Self {
-        TargetDifficultyCache {
-            inner: RwLock::new(TargetDifficultyCacheInner::new(capacity)),
-        }
-    }
-
-    pub fn get(&self, hash: &H256) -> Option<U256> {
-        let inner = self.inner.read();
-        inner.cache.get(hash).map(|diff| *diff)
-    }
-
-    pub fn set(&self, hash: H256, difficulty: U256) {
-        let mut inner = self.inner.write();
-        while inner.is_full() {
-            inner.evict_one();
-        }
-        inner.insert(hash, difficulty);
-    }
-}
-
-//FIXME: Add logic for persisting entries
-/// This is a data structure to cache the computed target difficulty
-/// of a adjustment period. Each element is indexed by the hash of
-/// the upper boundary block of the period.
-#[derive(DeriveMallocSizeOf)]
-pub struct TargetDifficultyManager {
-    cache: TargetDifficultyCache,
-}
-
-impl TargetDifficultyManager {
-    pub fn new(capacity: usize) -> Self {
-        TargetDifficultyManager {
-            cache: TargetDifficultyCache::new(capacity),
-        }
-    }
-
-    pub fn get(&self, hash: &H256) -> Option<U256> { self.cache.get(hash) }
-
-    pub fn set(&self, hash: H256, difficulty: U256) {
-        self.cache.set(hash, difficulty);
-    }
 }
 
 #[test]
